+++ lib/td_bg/user_search_filters.ex
@@ -43,7 +43,7 @@ defmodule TdBg.UserSearchFilters do
   defp maybe_filter(results, %{jti: jti}) do
     case Permissions.permitted_domain_ids(jti, "view_published_business_concepts") do
       [] ->
-        if Permissions.is_default_permission?("view_published_business_concepts"),
+        if Permissions.default_permission?("view_published_business_concepts"),
           do: results,
           else: []