+++ lib/td_bg_web/controllers/search_controller.ex
@@ -1,19 +1,11 @@
 defmodule TdBgWeb.SearchController do
   use TdBgWeb, :controller
   import Canada, only: [can?: 2]
-  use PhoenixSwagger
+
   alias TdBg.BusinessConcepts.BusinessConcept
   alias TdBgWeb.ErrorView
 
   alias TdBg.Search.Indexer
-
-  swagger_path :reindex_all do
-    description("Reindex all ES indexes with DB content")
-    produces("application/json")
-    response(202, "Accepted")
-    response(403, "Unauthorized")
-    response(500, "Client Error")
-  end
 
   def reindex_all(conn, _params) do
     claims = conn.assigns[:current_resource]