+++ lib/td_bg_web/endpoint.ex
@@ -1,12 +1,6 @@
 defmodule TdBgWeb.Endpoint do
   use Phoenix.Endpoint, otp_app: :td_bg
 
-  # Serve at "/" the static files from "priv/static" directory.
-  #
-  # You should set gzip to true if you are running phoenix.digest
-  # when deploying your static files in production.
-  plug Plug.Static, at: "/", from: :td_bg, gzip: false, only: ~w(swagger.json)
-
   # Code reloading can be explicitly enabled under the
   # :code_reloader configuration of your endpoint.
   if code_reloading? do
@@ -39,19 +33,4 @@ defmodule TdBgWeb.Endpoint do
   plug Corsica, origins: "*"
 
   plug TdBgWeb.Router
-
-  @doc """
-  Callback invoked for dynamically configuring the endpoint.
-
-  It receives the endpoint configuration and checks if
-  configuration should be loaded from the system environment.
-  """
-  def init(_key, config) do
-    if config[:load_from_system_env] do
-      port = System.get_env("PORT") || raise "expected the PORT environment variable to be set"
-      {:ok, Keyword.put(config, :http, [:inet6, port: port])}
-    else
-      {:ok, config}
-    end
-  end
 end